Closes #140 dblclick dont generate remove or lock anymore
